Election page has many faces. Move OngoingElection to component.
