+++ packages/webapp/src/elections/components/index.ts
@@ -1,2 +1,3 @@
 export * from "./pie-mer";
 export * from "./election-member-chips";
+export * from "./ongoing-election";
